# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScreenIQ Background Screening Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
    post:
      operationId: createCandidate
      tags: [candidates]
      summary: Create a candidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      summary: Get a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
        "404":
          description: Not found
    patch:
      operationId: updateCandidate
      tags: [candidates]
      summary: Update a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    delete:
      operationId: deleteCandidate
      tags: [candidates]
      summary: Delete a candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Screening ────────────────────────────────────────────────
...
  /candidates/{candidateId}/adverse-action:
    get:
      operationId: getAdverseAction
      tags: [adverse-action]
      summary: Get the adverse action record for a candidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdverseAction"
        "404":
          description: No adverse action on file
    post:
      operationId: createAdverseAction
      tags: [adverse-action]
      summary: Initiate adverse action workflow for a candidate
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdverseActionInput"
      responses:
        "201":
          description: Adverse action initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdverseAction"
    patch:
      operationId: updateAdverseAction
      tags: [adverse-action]
      summary: Advance adverse action stage
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdverseActionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdverseAction"
  /candidates/{candidateId}/adverse-action/notices:
    get:
      operationId: listAdverseActionNotices
      tags: [adverse-action]
      summary: List all notices for a candidate's adverse action
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdverseActionNotice"
    post:
      operationId: generateAdverseActionNotice
      tags: [adverse-action]
      summary: Generate a pre-adverse or final adverse action notice
      parameters:
        - name: candidateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoticeInput"
      responses:
        "201":
          description: Notice generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdverseActionNotice"

  # ── Disputes ─────────────────────────────────────────────────
...
components:
  schemas:
...
    Candidate:
      type: object
      required: [id, name, position, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        ssnLastFour:
          type: ["string", "null"]
        position:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed, flagged, adverse_action]
        portalToken:
          type: string
        latestRunId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CandidateInput:
      type: object
      required: [name, position]
      properties:
        name:
          type: string
          minLength: 1
        dateOfBirth:
          type: string
        ssnLastFour:
          type: string
          maxLength: 4
        position:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
    CandidateUpdate:
      type: object
      properties:
        name:
          type: string
        dateOfBirth:
          type: string
        ssnLastFour:
          type: string
        position:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string

    # ── Screening ─────────────────────────────────────────────
...
    AdverseActionInput:
      type: object
      required: [screeningRunId]
      properties:
        screeningRunId:
          type: integer
        reason:
          type: string
    AdverseActionUpdate:
      type: object
      properties:
        stage:
          type: string
          enum: [pre_adverse, waiting_period, final_adverse, closed]
    AdverseAction:
      type: object
      required: [id, candidateId, screeningRunId, stage, createdAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        screeningRunId:
          type: integer
        stage:
          type: string
          enum: [pre_adverse, waiting_period, final_adverse, closed]
        reason:
          type: ["string", "null"]
        preAdverseNoticeSentAt:
          type: ["string", "null"]
        waitingPeriodEndsAt:
          type: ["string", "null"]
        finalNoticeAt:
          type: ["string", "null"]
        closedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    NoticeInput:
      type: object
      required: [noticeType]
      properties:
        noticeType:
          type: string
          enum: [pre_adverse, final_adverse]
    AdverseActionNotice:
      type: object
      required: [id, adverseActionId, noticeType, content, createdAt]
      properties:
        id:
          type: integer
        adverseActionId:
          type: integer
        noticeType:
          type: string
          enum: [pre_adverse, final_adverse]
        content:
          type: string
        sentAt:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── Disputes ──────────────────────────────────────────────
...