# lib/api-zod/src/generated/types/noticeInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ScreenIQ Background Screening Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { NoticeInputNoticeType } from './noticeInputNoticeType';

export interface NoticeInput {
  noticeType: NoticeInputNoticeType;
}